+++ backend/src/test/kotlin/app/cliq/backend/shared/EmailServiceUnitTests.kt
@@ -1,7 +1,6 @@
-package app.cliq.backend.unit.service
+package app.cliq.backend.shared
 
 import app.cliq.backend.config.EmailProperties
-import app.cliq.backend.shared.EmailService
 import io.pebbletemplates.pebble.PebbleEngine
 import jakarta.mail.internet.MimeMessage
 import org.junit.jupiter.api.BeforeEach
@@ -20,7 +19,7 @@ import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 @ExtendWith(MockitoExtension::class)
-class EmailServiceTests {
+class EmailServiceUnitTests {
     @Mock
     private lateinit var emailProperties: EmailProperties
